+++ CHANGELOG.md
@@ -28,3 +28,7 @@
 1. 修改/metrics/dynamic下的进程列表返回内容，忽略一些空值
 2. 修正agent升级后agent版本号和go版本号的埋点数据问题
 3. 修改打包程序增加epoch
+
+## TODO
+
+1. API接口返回数据结构中增加extime字段